[TaskScheduler][TO BE TESTED] Added makeTask method and removed pointer array and dynamic allocation.

+++ src/shared/scheduler/TaskSchedulerData.h
@@ -43,7 +43,7 @@ namespace Boardcore
  */
 struct TaskStatsResult
 {
-    uint8_t id;
+    size_t id;
     uint32_t period;
     StatsResult activationStats;
     StatsResult periodStats;
